+++ src/pages/Board/BoardModifyPage.js
@@ -1,4 +1,4 @@
-import "../style/Board.css";
+import "./Board.css";
 import {
   Dropdown,
   InputGroup,
@@ -7,12 +7,13 @@ import {
   Form,
   FloatingLabel,
 } from "react-bootstrap";
-function BoardWritePage() {
+
+function BoardModifyPage() {
   return (
-    <div className="BoardWritePage">
+    <div className="BoardModifyPage">
       <div className="row">
         <div className="col-md-9">
-          <h1>게시글 작성</h1>
+          <h1>게시글 수정</h1>
         </div>
         <div className="col-md-3">
           <button className="boardButton">꾸미기 설정</button>
@@ -62,4 +63,4 @@ function BoardWritePage() {
   );
 }
 
-export default BoardWritePage;
+export default BoardModifyPage;
